+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## 2023-12-7
 
 - Initial beta release.
+
+## 2023-12-18
+
+- Added proof deleted webhook, with a simplified proof payload. Only containing the proof id.